backend/requirements.txt
fastapi
uvicorn
requests
python-chess
